+++ src/main/java/net/frozenblock/wilderwild/mixin/entity/stray/StrayMixin.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2023-2024 FrozenBlock
+ * This file is part of Wilder Wild.
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, see <https://www.gnu.org/licenses/>.
+ */
+
+package net.frozenblock.wilderwild.mixin.entity.stray;
+
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import net.frozenblock.wilderwild.tag.WilderBiomeTags;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.entity.monster.Stray;
+import net.minecraft.world.level.ServerLevelAccessor;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(Stray.class)
+public class StrayMixin {
+
+	@WrapOperation(
+		method = "checkStraySpawnRules",
+		at = @At(
+			value = "INVOKE",
+			target = "Lnet/minecraft/world/level/ServerLevelAccessor;canSeeSky(Lnet/minecraft/core/BlockPos;)Z"
+		)
+	)
+	private static boolean wilderWild$spawnInFrozenCaves(ServerLevelAccessor instance, BlockPos pos, Operation<Boolean> original) {
+		return original.call(instance, pos) || instance.getBiome(pos).is(WilderBiomeTags.STRAYS_CAN_SPAWN_UNDERGROUND);
+	}
+
+}