+++ src/main/java/net/frozenblock/wilderwild/registry/RegisterWorldgen.java
@@ -29,6 +29,7 @@ import net.frozenblock.wilderwild.world.biome.DarkTaiga;
 import net.frozenblock.wilderwild.world.biome.DyingForest;
 import net.frozenblock.wilderwild.world.biome.DyingMixedForest;
 import net.frozenblock.wilderwild.world.biome.FlowerField;
+import net.frozenblock.wilderwild.world.biome.FrozenCaves;
 import net.frozenblock.wilderwild.world.biome.JellyfishCaves;
 import net.frozenblock.wilderwild.world.biome.MagmaticCaves;
 import net.frozenblock.wilderwild.world.biome.MixedForest;
@@ -57,6 +58,7 @@ public final class RegisterWorldgen {
 	public static final ResourceKey<Biome> WARM_RIVER = WarmRiver.INSTANCE.getKey();
 	public static final ResourceKey<Biome> WARM_BEACH = WarmBeach.INSTANCE.getKey();
 	// Cave Biomes
+	public static final ResourceKey<Biome> FROZEN_CAVES = FrozenCaves.INSTANCE.getKey();
 	public static final ResourceKey<Biome> JELLYFISH_CAVES = JellyfishCaves.INSTANCE.getKey();
 	public static final ResourceKey<Biome> MAGMATIC_CAVES = MagmaticCaves.INSTANCE.getKey();
 	// Transition Biomes
@@ -108,6 +110,7 @@ public final class RegisterWorldgen {
 		register(context, WARM_RIVER, WarmRiver.INSTANCE.create(context));
 		register(context, WARM_BEACH, WarmBeach.INSTANCE.create(context));
 		// CAVE BIOMES
+		register(context, FROZEN_CAVES, FrozenCaves.INSTANCE.create(context));
 		register(context, JELLYFISH_CAVES, JellyfishCaves.INSTANCE.create(context));
 		register(context, MAGMATIC_CAVES, MagmaticCaves.INSTANCE.create(context));
 		// TRANSITION BIOMES